requirements.txt
APScheduler==3.6.3
Authlib==0.14.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Flask==1.1.1
google-api-python-client==1.7.12
google-auth==1.11.2
google-auth-httplib2==0.0.3
gspread==3.2.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
python-gflags==3.1.2
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==4.0
six==1.14.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
